backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
langchain-huggingface
faiss-cpu
pypdf
python-dotenv
sentence-transformers
pydantic
python-multipart
youtube-transcript-api
beautifulsoup4
pytube
lxml
